implement like and dislike, implement if film exist cant be post (#5)
Marcell

+++ showcase/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render
 from django.http import HttpResponse
+from rest_framework import serializers
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from .models import Film
@@ -10,25 +11,45 @@ class FilmList(APIView):
     def get(self,request):
         #TODO Implement sort by like, implement sort descending and ascending released year
         film = Film.objects.all()
+       
         serializers = FilmAllSerializer(film, many=True)
+
         return Response({"status" : 200,
             "message" : "Success",
             "data":serializers.data})
 
     def post(self,request):
-        film = Film.objects.create(
-            title = request.data['title'],
-            imageUrl = request.data['imageUrl'],
-            trailerUrl = request.data['trailerUrl'],
-            like=0,
-            genre = request.data['genre'],
-            released_date = request.data['released_date'],
-        )
 
-        serializers = FilmAllSerializer(film) 
-        return Response({"status" : 201,
-            "message" : "Created",
-            "data":serializers.data})
+        film = Film.objects.all()      
+        serializers = FilmAllSerializer(film, many=True)
+        film_exist = False
+        for data in serializers.data:
+            if request.data['title'] == data['title'] and request.data['released_year'] == data['released_year']:
+                film_exist=True
+                break
+            else:
+                film_exist=False
+            
+        if not film_exist:
+            film = Film.objects.create(
+                title = request.data['title'],
+                imageUrl = request.data['imageUrl'],
+                trailerUrl = request.data['trailerUrl'],
+                genre = request.data['genre'],
+                released_year = request.data['released_year'],
+            )
+            serializers = FilmDetailSerializer(film)
+            return Response({"status" : 201,
+        "message" : "Created",
+        "data":serializers.data})
+
+        else:
+            return Response({
+                "status" : 204,
+        "message" : "Success no data, existed film"})
+
+
+        
 
 class FilmDetail(APIView):
     def get(self, request, id):
@@ -46,37 +67,47 @@ class FilmDetail(APIView):
             "data" : serializers.data
         })
 
-    def put(self, request, id):
+    def put(self, request, id,action):
         try :
-            pass
-            ## hanya contoh
-            # TODO: Implement tambah like 
-            # film = Film.objects.get(id=id)
-            # film.name = request.data['name']
-            # film.author = request.data['author']
-            # film.volume = request.data['volume']
-            # film.total_page = request.data['total_page']
-            # film.publisher = request.data['publisher']
-            # film.published_date = request.data['published_date']
-            # film.save()
+            film = Film.objects.get(id=id)
+            if action == "like":
+                film.like += 1
+            elif action == "dislike":
+                film.dislike += 1
+            else:
+                return Response({
+                    "error": "hanya bisa like atau dislike"
+                })
+            film.save()
         except Film.DoesNotExist:
              return Response({
-                "error" : "buku tidak ditemukan"
+                "error" : "film tidak ditemukan"
             })
 
-        # serializers = FilmDetailSerializer(film)
-        # return Response({
-        #     "status" : 201,
-        #     "message" : "update successfull",
-        #     "data" : serializers.data
-        # })
-
-
-    def delete(self, request, id) :
-        film = Film.objects.get(id=id)
-        film.delete()
-
+        serializers = FilmDetailSerializer(film)
         return Response({
             "status" : 201,
-            "message" : "delete success",
+            "message" : "update successfull",
+            "data" : serializers.data
         })
+
+
+    def delete(self, request, id,action) :
+        try:
+            if action == "delete":
+                film = Film.objects.get(id=id)
+                film.delete()
+
+                return Response({
+                    "status" : 201,
+                    "message" : "delete success",
+                })
+            else:
+                return Response({
+                    "status" : 204,
+                    "message" : "/delete is required",
+                })
+        except Film.DoesNotExist:
+            return Response({
+                "error" : "film tidak ditemukan"
+            })
